feat: Add empty forms component to display when there are no forms
This commit adds a new component called "EmptyForms" to display when there are no forms in the forms list. The component includes an image, a message, and a button to create a new form. It is used in the "FormsList" component when the forms array is empty and the data has been fetched.