feat: Add "EmptyForms" component to display when no forms exist
This commit introduces a new component called "EmptyForms" that is used to display a message, image, and a button to create a new form when there are no forms in the forms list. The component is utilized in the "FormsList" component when the forms array is empty and the data has been fetched.